requirements.txt
requests
sqlalchemy
psycopg2-binary
python-dotenv
